backend1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdf2docx==0.5.6
python-docx==0.8.11
PyMuPDF==1.23.7
aiofiles==23.2.1
pydub==0.25.1
scipy==1.11.4
numpy==1.26.2
moviepy==1.0.3
SpeechRecognition==3.10.0
google-api-python-client==2.108.0
google-cloud-speech==2.21.0
gTTS==2.3.2
ffmpeg-python==0.2.0
sounddevice==0.4.6
soundfile==0.12.1
chardet==5.2.0
reportlab==4.0.8
striprtf==0.0.26
pytesseract==0.3.10
Pillow==10.1.0 